code/requirements.txt
# CP1 Production Dependencies
# ============================

# Core ML
xgboost>=2.0.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.3.0

# Bitcoin
python-bitcoinrpc>=1.0
pyzmq>=25.0.0

# Monitoring
prometheus_client>=0.17.0

# Caching & Message Bus
redis>=5.0.0

# Explainability
shap>=0.42.0

# Testing
pytest>=8.0.0
pytest-cov>=4.0.0

# HTTP/API (for health checks)
requests>=2.31.0
